shared/src/main/java/com/google/samples/apps/iosched/shared/domain/auth/ObserveUserAuthStateUseCase.kt
/*
 * Copyright 2018 Google LLC
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     https://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package com.google.samples.apps.iosched.shared.domain.auth

import com.google.samples.apps.iosched.shared.data.signin.AuthenticatedUserInfo
import com.google.samples.apps.iosched.shared.data.signin.datasources.AuthStateUserDataSource
import com.google.samples.apps.iosched.shared.di.DefaultDispatcher
import com.google.samples.apps.iosched.shared.domain.FlowUseCase
import com.google.samples.apps.iosched.shared.result.Result
import kotlinx.coroutines.CoroutineDispatcher
import kotlinx.coroutines.flow.Flow
import kotlinx.coroutines.flow.map
import javax.inject.Inject
import javax.inject.Singleton

/**
 * A [FlowUseCase] that observes a data source to generate an [AuthenticatedUserInfo].
 *
 * [AuthStateUserDataSource] provides general user information, like user IDs.
 */
@Singleton
open class ObserveUserAuthStateUseCase @Inject constructor(
    private val authStateUserDataSource: AuthStateUserDataSource,
    @DefaultDispatcher defaultDispatcher: CoroutineDispatcher
) : FlowUseCase<Any, AuthenticatedUserInfo?>(defaultDispatcher) {

    override fun execute(parameters: Any): Flow<Result<AuthenticatedUserInfo?>> =
        authStateUserDataSource.getBasicUserInfo().map { userResult ->
            if (userResult is Result.Success) {
                Result.Success(userResult.data)
            } else {
                Result.Error(Exception("FirebaseAuth error"))
            }
        }
}
